\name{adevent}
\alias{adevent}
\docType{data}
\title{
Adverse Events Data
}
\description{
Indicators of 28 adverse events in a two-arm clinical trial.
}
\usage{data(adevent)}
\format{
  A data frame with 160 observations on the following 29 variables.
  \describe{
    \item{\code{E1}}{a factor with levels \code{no event} \code{event}}
    \item{\code{E2}}{a factor with levels \code{no event} \code{event}}
    \item{\code{E3}}{a factor with levels \code{no event} \code{event}}
    \item{\code{E4}}{a factor with levels \code{no event} \code{event}}
    \item{\code{E5}}{a factor with levels \code{no event} \code{event}}
    \item{\code{E6}}{a factor with levels \code{no event} \code{event}}
    \item{\code{E7}}{a factor with levels \code{no event} \code{event}}
    \item{\code{E8}}{a factor with levels \code{no event} \code{event}}
    \item{\code{E9}}{a factor with levels \code{no event} \code{event}}
    \item{\code{E10}}{a factor with levels \code{no event} \code{event}}
    \item{\code{E11}}{a factor with levels \code{no event} \code{event}}
    \item{\code{E12}}{a factor with levels \code{no event} \code{event}}
    \item{\code{E13}}{a factor with levels \code{no event} \code{event}}
    \item{\code{E14}}{a factor with levels \code{no event} \code{event}}
    \item{\code{E15}}{a factor with levels \code{no event} \code{event}}
    \item{\code{E16}}{a factor with levels \code{no event} \code{event}}
    \item{\code{E17}}{a factor with levels \code{no event} \code{event}}
    \item{\code{E18}}{a factor with levels \code{no event} \code{event}}
    \item{\code{E19}}{a factor with levels \code{no event} \code{event}}
    \item{\code{E20}}{a factor with levels \code{no event} \code{event}}
    \item{\code{E21}}{a factor with levels \code{no event} \code{event}}
    \item{\code{E22}}{a factor with levels \code{no event} \code{event}}
    \item{\code{E23}}{a factor with levels \code{no event} \code{event}}
    \item{\code{E24}}{a factor with levels \code{no event} \code{event}}
    \item{\code{E25}}{a factor with levels \code{no event} \code{event}}
    \item{\code{E26}}{a factor with levels \code{no event} \code{event}}
    \item{\code{E27}}{a factor with levels \code{no event} \code{event}}
    \item{\code{E28}}{a factor with levels \code{no event} \code{event}}
    \item{\code{group}}{group indicator.}
  }
}
\details{
The data is provided by \bibcitet{multcomp::westfall1999}, page 242, and contains binary
indicators of 28 adverse events (\code{E1},..., \code{E28}) for
two arms (\code{group}).
}
\references{\bibshow{*}}
\keyword{datasets}
