% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Symbol.R
\docType{methods}
\name{as.character,Symbol-method}
\alias{as.character,Symbol-method}
\title{Get the name of the Symbol instance}
\usage{
\S4method{as.character}{Symbol}(x)
}
\arguments{
\item{x}{An instance derived from the Symbol}
}
\value{
Name of object as a character string (same as name method)
}
\description{
Get the name of the Symbol instance
}

