% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/expressions.R
\name{chrlinfct2matrix}
\alias{chrlinfct2matrix}
\title{Interpreter for character representations of linear functions}
\usage{
chrlinfct2matrix(ex, var, verbose = F, zerocheck = T, ...)
}
\arguments{
\item{ex}{An (optionally named) character vector of one or more null hypotheses}

\item{var}{Character vector of effect names}

\item{verbose}{Trace the interpreter actions if true}

\item{zerocheck}{Enable checking for zero or near zero coefficients.}

\item{...}{passed down from ghlt, unused}
}
\value{
list(K, m, alternative, h0 = list( labels = list( original, generated ))) with
         \itemize{
              \item K            Matrix of numeric effect coefficients, with \code{rownames(K)}
                                 set to the character string representing the left-hand side
              \item m            Vector of right hand side values
              \item alternative  One of: 'less', 'greater' or 'two.sided'
              \item h0           A list describing the hypotheses passed in argument `ex`
                  \itemize{
                       \item lhs          Array of character strings representing the left hand side
                       \item rhs          Array of character strings representing the right hand side
                       \item labels       Labels for the list of equations argument passed in argument ex
                            \itemize{
                                 \item original   Names passed along with argument ex (may be null)
                                 \item generated  Names like 'h0.1', 'h0.2', ...      (null if original was not null )
                  }
              }
        }
}
\description{
Extract the effect coefficients from a potentially named list of null hypotheses
}
\examples{
a <- '1/2*( x1 -   x2 ) +1 = 1/3 * x1:x2'
b <- '1/3*( x2 - 2*x1 ) -3 = 2 + exp((-2)^3) * x1:x2'
multcomp:::chrlinfct2matrix(ex = c('test x' = a,  'test y' = b), var = c('x1', 'x2', 'x1:x2'))
}
\seealso{
expression2coef
}

