% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Symbol.R
\docType{methods}
\name{coef<-,Symbol-method}
\alias{coef<-,Symbol-method}
\title{Sets the coefficient of the Symbol instance to a new value}
\usage{
\S4method{coef}{Symbol}(object) <- value
}
\arguments{
\item{object}{Instance of class Symbol}

\item{value}{A scalar}
}
\value{
Instance of object
}
\description{
Sets the coefficient of the Symbol instance to a new value
}

