% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Symbol.R
\name{coef<-}
\alias{coef<-}
\title{Generic method  coef<-}
\usage{
coef(object) <- value
}
\arguments{
\item{object}{Any object}

\item{value}{A scalar}
}
\value{
Instance of object
}
\description{
Generic method  coef<-
}

