% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Symbol.R
\docType{methods}
\name{coef,Symbol-method}
\alias{coef,Symbol-method}
\title{Get the value of the coefficient associated with the symbol}
\usage{
\S4method{coef}{Symbol}(object, ...)
}
\arguments{
\item{object}{An instance derived from class Symbol}

\item{...}{Other arguments prescribed by \code{\link[stats]{coef}}, ignored}
}
\value{
Value of the coefficient
}
\description{
Get the value of the coefficient associated with the symbol
}

