% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Symbol.R
\name{coef}
\alias{coef}
\title{Generic method coef}
\usage{
coef(object, ...)
}
\arguments{
\item{object}{Any object}

\item{...}{Other arguments prescribed by \code{\link[stats]{coef}}, ignored}
}
\value{
Coefficient value
}
\description{
Generic method coef
}

