% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Effect.R
\docType{class}
\name{Effect-class}
\alias{Effect-class}
\title{Represent an effect symbol as defined by the model, extends Symbol}
\description{
The instance tracks the name and the coefficient of a
       single occurence of an effect symbol.
       In \eqn{ 3*(x2 - x1) - 4*(x3 -x1) == 0} the same effect
       x1 gets referenced multiple times. This can only be supported
       if the two instances do not share the same memory region,
       as it is the case with symbol nodes returned from \code{parse()}
}

