% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Effect.R
\name{Effect}
\alias{Effect}
\title{Effect constructor}
\usage{
Effect(name, coef = 1)
}
\arguments{
\item{name}{Effect name}

\item{coef}{Effect coefficient}
}
\value{
A new instance of class Effect
}
\description{
Effect constructor
}

