% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Symbol.R
\docType{methods}
\name{==,Symbol,character-method}
\alias{==,Symbol,character-method}
\title{Compares the name of an instance derived from class Symbol against a character string}
\usage{
\S4method{==}{Symbol,character}(e1, e2)
}
\arguments{
\item{e1}{Instance of Symbol}

\item{e2}{A character string}
}
\value{
True if name for instance is equal to string e2
}
\description{
Compares the name of an instance derived from class Symbol against a character string
}

