% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Symbol.R
\docType{methods}
\name{==,Symbol,Symbol-method}
\alias{==,Symbol,Symbol-method}
\title{Compare the name of a Symbol instance against the name of another instance of class Symbol}
\usage{
\S4method{==}{Symbol,Symbol}(e1, e2)
}
\arguments{
\item{e1}{Instance of Symbol}

\item{e2}{Instance of Symbol}
}
\value{
true if both names are equal
}
\description{
Compare the name of a Symbol instance against the name of another instance of class Symbol
}

