% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Literal.R
\docType{methods}
\name{initialize,Literal-method}
\alias{initialize,Literal-method}
\title{Initialize method for instances of class Literal}
\usage{
\S4method{initialize}{Literal}(.Object, value)
}
\arguments{
\item{.Object}{Instance derived from class Literal}

\item{value}{Numeric value, also serving as symbol name}
}
\value{
object instance
}
\description{
Initialize method for instances of class Literal
}

