% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Symbol.R
\docType{methods}
\name{initialize,Symbol-method}
\alias{initialize,Symbol-method}
\title{Initialize method for instances of class Symbol}
\usage{
\S4method{initialize}{Symbol}(.Object, name, coef)
}
\arguments{
\item{.Object}{instance of class Symbol}

\item{name}{Symbol name}

\item{coef}{Symbol coefficient}
}
\value{
An instance of Symbol or NULL if initialization fails
}
\description{
Initialize method for instances of class Symbol
}

