% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Effect.R
\docType{methods}
\name{is.Effect,ANY-method}
\alias{is.Effect,ANY-method}
\title{Tests if an object is an instance of class Effect}
\usage{
\S4method{is.Effect}{ANY}(object)
}
\arguments{
\item{object}{Any object (except an instance of class Effect)}
}
\value{
Always false
}
\description{
Tests if an object is an instance of class Effect
}

