% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Effect.R
\docType{methods}
\name{is.Effect,Effect-method}
\alias{is.Effect,Effect-method}
\title{Tests if an object is an instance of class Effect}
\usage{
\S4method{is.Effect}{Effect}(object)
}
\arguments{
\item{object}{Instance of class Effect}
}
\value{
Always true
}
\description{
Tests if an object is an instance of class Effect
}

