% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Effect.R
\name{is.Effect}
\alias{is.Effect}
\title{Generic method is.Effect}
\usage{
is.Effect(object)
}
\arguments{
\item{object}{any object}
}
\value{
True if the instance is of class Effect
}
\description{
Generic method is.Effect
}

