% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Literal.R
\docType{methods}
\name{is.Literal,ANY-method}
\alias{is.Literal,ANY-method}
\title{Tests if object is an instance of class Literal}
\usage{
\S4method{is.Literal}{ANY}(object)
}
\arguments{
\item{object}{Any object (except an instance derived from Literal)}
}
\value{
Always false
}
\description{
Tests if object is an instance of class Literal
}

