% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Literal.R
\docType{methods}
\name{is.Literal,Literal-method}
\alias{is.Literal,Literal-method}
\title{Tests if object is an instance of class Literal}
\usage{
\S4method{is.Literal}{Literal}(object)
}
\arguments{
\item{object}{An instance derived from Literal}
}
\value{
Always true
}
\description{
Tests if object is an instance of class Literal
}

