% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Literal.R
\name{is.Literal}
\alias{is.Literal}
\title{Generic method Is.Effect}
\usage{
is.Literal(object)
}
\arguments{
\item{object}{Any object}
}
\value{
True if object is an instance of class Effect
}
\description{
Generic method Is.Effect
}

