% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Symbol.R
\docType{methods}
\name{is.Symbol,ANY-method}
\alias{is.Symbol,ANY-method}
\title{Test if object is an instance derived from  class Symbol}
\usage{
\S4method{is.Symbol}{ANY}(object)
}
\arguments{
\item{object}{Any object (except an instance derived from class Symbol)}
}
\value{
Always false
}
\description{
Test if object is an instance derived from  class Symbol
}

