% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Symbol.R
\docType{methods}
\name{is.Symbol,Symbol-method}
\alias{is.Symbol,Symbol-method}
\title{Test if object is an instance derived from  class Symbol}
\usage{
\S4method{is.Symbol}{Symbol}(object)
}
\arguments{
\item{object}{An instance derived from class Symbol}
}
\value{
Always true
}
\description{
Test if object is an instance derived from  class Symbol
}

