% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Symbol.R
\name{is.Symbol}
\alias{is.Symbol}
\title{Generic method is.Symbol}
\usage{
is.Symbol(object)
}
\arguments{
\item{object}{Any object}
}
\value{
True if object is an instance derived from class Symbol
}
\description{
Generic method is.Symbol
}

