% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Literal.R
\name{Literal}
\alias{Literal}
\title{Literal constructor}
\usage{
Literal(value)
}
\arguments{
\item{value}{Numeric value, also serving as symbol name}
}
\value{
A new instance of class Literal
}
\description{
Literal constructor
}

