\name{multcomp-internal}
\alias{ct}
\alias{getdigits}
\alias{nicepaste}
\alias{parseformula}
\title{Internal multcomp functions}
\description{
 Internal multcomp functions.
}
\usage{
ct(x, contrasts=FALSE)
getdigits(x)
nicepaste(x, pattern)
parseformula(formula, data, subset, na.action, whichf, ...)
}

}
\details{
  This functions are not to be called by the user.
}
\keyword{internal}
