% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Symbol.R
\docType{methods}
\name{name,Symbol-method}
\alias{name,Symbol-method}
\title{Get the name of the Symbol instance}
\usage{
\S4method{name}{Symbol}(object)
}
\arguments{
\item{object}{An instance derived from class Symbol}
}
\value{
Name of object as a character string
}
\description{
Get the name of the Symbol instance
}

