\name{print.hmtest}
\alias{print.hmtest}
\title{Print Method for hmtest Object}
\description{
  Print object of class \code{"hmtest"} in nice layout.
}
\usage{
\method{print}{hmtest}(x, digits, \dots)
}
\arguments{
  \item{x}{Object of class \code{"hmtest"}.}
  \item{digits}{how many digits should be printed.}
  \item{\dots}{further arguments to be passed to or from methods.}  
}
\details{
  Prints a \code{hmtest} object.
}
\value{
  none
}
}
\keyword{htest}
