\name{print.hmtestp}
\alias{print.hmtestp}
\title{Print Method for hmtestp Object}
\description{
  Print object of class \code{"hmtestp"} in nice layout.
}
\usage{
\method{print}{hmtestp}(x, digits, \dots)
}
\arguments{
  \item{x}{Object of class \code{"hmtestp"}.}
  \item{digits}{how many digits should be printed.}
  \item{\dots}{further arguments to be passed to or from methods.}  
}
\details{
  Print a \code{hmtestp} object. 
}
\value{
  none
}
\keyword{htest}
