% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Symbol.R
\docType{methods}
\name{quoted,Symbol-method}
\alias{quoted,Symbol-method}
\title{Quote the name of the Symbol instance using backticks ('`')}
\usage{
\S4method{quoted}{Symbol}(object)
}
\arguments{
\item{object}{An instance derived from class Symbol}
}
\value{
Name of object as a quoted character string
}
\description{
Quote the name of the Symbol instance using backticks ('`')
}

