% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Symbol.R
\name{quoted}
\alias{quoted}
\title{Quote the name of the Symbol instance using backticks ('`')}
\usage{
quoted(object)
}
\arguments{
\item{object}{Any object}
}
\value{
Name of object as a quoted character string
}
\description{
Quote the name of the Symbol instance using backticks ('`')
}

