\name{sbp}
\alias{sbp}
\docType{data}
\title{
  Systolic Blood Pressure Data
}
\description{
  Systolic blood pressure, age and gender of 69 people.
}
\usage{data("sbp")}
\format{
  A data frame with 69 observations on the following 3 variables.
  \describe{
    \item{\code{gender}}{a factor with levels \code{male} \code{female}}
    \item{\code{sbp}}{systolic blood pressure in mmHg}
    \item{\code{age}}{age in years}
  }
}
\details{
Data were obtained from \bibcitet{multcomp::kleinbaum1998}.
}
\references{\bibshow{*}}
\keyword{datasets}
