\name{summary.hmtest}
\alias{summary.hmtest}
\alias{print.summary.hmtest}
\title{Summarizing Simultaneous Confidence Intervals}
\description{
  \code{summary} method for class \code{hmtest}.
}
\usage{
\method{summary}{hmtest}(object, \dots)
}
\arguments{
  \item{object}{Object of class \code{"hmtest"}.}
  \item{\dots}{further arguments to be passed to or from methods.}  
}
\details{
  Summarize all information from an \code{hmtest} object.
}
\value{
  none
}
}
\keyword{htest}
