\name{summary.hmtestp}
\alias{summary.hmtestp}
\alias{print.summary.hmtestp}
\title{Summarizing Simultaneous Confidence Intervals}
\description{
  \code{summary} method for class \code{hmtestp}.
}
\usage{
\method{summary}{hmtestp}(object, \dots)
}
\arguments{
  \item{object}{Object of class \code{"hmtestp"}.}
  \item{\dots}{further arguments to be passed to or from methods.}  
}
\details{
  Summarize all information from an \code{hmtestp} object.
}
\value{
  none
}
}
\keyword{htest}
