% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Symbol.R
\docType{class}
\name{Symbol-class}
\alias{Symbol-class}
\title{Represent an instance of an external symbol, e.g. `pi`}
\description{
Represent an instance of an external symbol, e.g. `pi`
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of the symbol}

\item{\code{coef}}{Coefficient of the symbol}
}}

