% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Symbol.R
\name{Symbol}
\alias{Symbol}
\title{Symbol constructor}
\usage{
Symbol(name, coef = 1)
}
\arguments{
\item{name}{Name of the Symbol}

\item{coef}{Symbol coefficient}
}
\value{
A new instance of class Symbol
}
\description{
Symbol constructor
}

